Move autocomplete component out of logic as a individual component
