+++ src/main/java/seedu/address/autocomplete/parser/AutoCompleteTagParser.java
@@ -1,11 +1,11 @@
-package seedu.address.logic.autocomplete.parser;
+package seedu.address.autocomplete.parser;
 
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import java.util.LinkedList;
 import java.util.List;
 
-import seedu.address.logic.autocomplete.AutoCompleteUtils;
+import seedu.address.autocomplete.AutoCompleteUtils;
 import seedu.address.model.Model;
 
 //@@author john19950730
