+++ src/main/java/seedu/address/logic/Logic.java
@@ -25,10 +25,4 @@ public interface Logic {
     /** Returns the list of input entered by the user, encapsulated in a {@code ListElementPointer} object */
     ListElementPointer getHistorySnapshot();
 
-    //@@author john19950730
-    /** Returns the list of commands that autocomplete function should complete for the user */
-    ListElementPointer getAutoCompleteSnapshot();
-
-    void updateAutoCompletePossibilities(String stub);
-
 }
