+++ src/test/java/seedu/address/autocomplete/parser/AutoCompleteCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.autocomplete.parser;
+package seedu.address.autocomplete.parser;
 
 import static org.junit.Assert.assertEquals;
