+++ src/test/java/seedu/address/autocomplete/parser/AutoCompleteWordInNameParserTest.java
@@ -1,8 +1,9 @@
-package seedu.address.logic.autocomplete.parser;
+package seedu.address.autocomplete.parser;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 import static seedu.address.testutil.TypicalPersons.ALICE;
+import static seedu.address.testutil.TypicalPersons.AMY;
 import static seedu.address.testutil.TypicalPersons.BENSON;
 import static seedu.address.testutil.TypicalPersons.CARL;
 import static seedu.address.testutil.TypicalPersons.DANIEL;
@@ -26,17 +27,16 @@ import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
-import seedu.address.model.tag.Tag;
 
-public class AutoCompleteTagParserTest {
+public class AutoCompleteWordInNameParserTest {
 
     @Rule
     public ExpectedException thrown = ExpectedException.none();
 
-    private AutoCompleteTagParser parser;
+    private AutoCompleteWordInNameParser parser;
     private ModelStubWithRequiredMethods mockModel;
     private final List<ReadOnlyPerson> allPersonsAdded = Arrays.asList(
-            new ReadOnlyPerson[]{ALICE, BENSON, CARL, DANIEL, ELLE, FIONA, GEORGE});
+            new ReadOnlyPerson[]{ALICE, AMY, BENSON, CARL, DANIEL, ELLE, FIONA, GEORGE});
 
     @Before
     public void fillMockModel() {
@@ -46,27 +46,33 @@ public class AutoCompleteTagParserTest {
         } catch (DuplicatePersonException ex) {
             fail("This exception should not be thrown.");
         }
-        parser = new AutoCompleteTagParser(mockModel);
+        parser = new AutoCompleteWordInNameParser(mockModel);
     }
 
     //@@author john19950730
     @Test
     public void testParsePossibilities() {
         // multiple matches
-        String preamble = "findtag f";
+        String preamble = "find A";
         assertEquals(parser.parseForPossibilities(preamble),
-                Arrays.asList(preamble + "riends",
-                        preamble + "amily",
+                Arrays.asList(preamble + "lice",
+                        preamble + "my",
                         preamble));
 
         // single match
-        preamble = "removetag c";
+        preamble = "find P";
         assertEquals(parser.parseForPossibilities(preamble),
-                Arrays.asList(preamble + "olleagues",
+                Arrays.asList(preamble + "auline",
+                        preamble));
+
+        // single duplicate match
+        preamble = "find Mey";
+        assertEquals(parser.parseForPossibilities(preamble),
+                Arrays.asList(preamble + "er",
                         preamble));
 
         // no match
-        preamble = "removetag friends g";
+        preamble = "find Z";
         assertEquals(parser.parseForPossibilities(preamble),
                 Arrays.asList(preamble));
     }
@@ -97,19 +103,9 @@ public class AutoCompleteTagParserTest {
         }
 
         @Override
-        public List<String> getAllTagsInAddressBook() {
-            //generate a unique tag list first
-            final ArrayList<Tag> tagsList = new ArrayList<Tag>();
-            for (ReadOnlyPerson person : personsAdded) {
-                for (Tag tag : person.getTags()) {
-                    if (tagsList.indexOf(tag) == -1) {
-                        tagsList.add(tag);
-                    }
-                }
-            }
-
-            return tagsList.stream()
-                    .map(tag -> tag.toString().substring(1, tag.toString().length() - 1))
+        public List<String> getAllNamesInAddressBook() {
+            return personsAdded.stream()
+                    .map(person -> person.getName().toString())
                     .collect(Collectors.toList());
         }
